pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "joblib>=1.4.2",
    "numpy>=2.2.2",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.42.0",
]
